fuzz: add missing CreateNonMembershipProof bin
While at it, also rename internal to internal_, so oss-fuzz can build
the package. Otherwise, usage of internal package is disallowed by go
compiler.

Updates #7921
