+++ fuzz/oss-fuzz-build.sh
@@ -6,7 +6,7 @@ compile_go_fuzzer "$FUZZ_ROOT"/fuzz/crypto/hd/deriveprivatekeyforpath Fuzz fuzz_
 compile_go_fuzzer "$FUZZ_ROOT"/fuzz/crypto/hd/newparamsfrompath Fuzz fuzz_crypto_hd_newparamsfrompath fuzz
 compile_go_fuzzer "$FUZZ_ROOT"/fuzz/crypto/types/compactbitarray/marshalunmarshal Fuzz fuzz_crypto_types_compactbitarray_marshalunmarshal fuzz
 
-compile_go_fuzzer "$FUZZ_ROOT"/fuzz/store/internal/proofs/createnonmembershipproof Fuzz fuzz_store_internal_proofs_createnonmembershipproof fuzz
+compile_go_fuzzer "$FUZZ_ROOT"/fuzz/store/internal_/proofs/createnonmembershipproof Fuzz fuzz_store_internal_proofs_createnonmembershipproof fuzz
 
 compile_go_fuzzer "$FUZZ_ROOT"/fuzz/tendermint/amino/decodetime Fuzz fuzz_tendermint_amino_decodetime fuzz
