--- fuzz/store/internal/proofs/createnonmembershipproof/corpus/1.txt
@@ -1,1 +1,0 @@
-{"Data":{"0rwn2fTHD73elajO7XCL":"dmFsdWVfZm9yXzByd24yZlRIRDczZWxhak83WENM","1Qal6alNMlhWUq7ULwOK":"dmFsdWVfZm9yXzFRYWw2YWxOTWxoV1VxN1VMd09L","1neAuLUhEEKdNmFKU4ra":"dmFsdWVfZm9yXzFuZUF1TFVoRUVLZE5tRktVNHJh","2R5jLba8JPgNEYQpWXqM":"dmFsdWVfZm9yXzJSNWpMYmE4SlBnTkVZUXBXWHFN","2xIn0e5U6iIGzmrNjXOL":"dmFsdWVfZm9yXzJ4SW4wZTVVNmlJR3ptck5qWE9M","4Nx12nTcGIGfO3HIXxwM":"dmFsdWVfZm9yXzROeDEyblRjR0lHZk8zSElYeHdN","4glYJRD3aR6OgL5co0Ju":"dmFsdWVfZm9yXzRnbFlKUkQzYVI2T2dMNWNvMEp1","5fCcGRr1thmESIBHYcOW":"dmFsdWVfZm9yXzVmQ2NHUnIxdGhtRVNJQkhZY09X","5vzeAhhI9oWG7HdZdIEH":"dmFsdWVfZm9yXzV2emVBaGhJOW9XRzdIZFpkSUVI","6Ik5ONVHpfq7bBx9Qlyo":"dmFsdWVfZm9yXzZJazVPTlZIcGZxN2JCeDlRbHlv","6OcjKZugrPPWzhDIDqZZ":"dmFsdWVfZm9yXzZPY2pLWnVnclBQV3poRElEcVpa","86KGslA97XUkoPbqu7Qi":"dmFsdWVfZm9yXzg2S0dzbEE5N1hVa29QYnF1N1Fp","8BdM7Hz1pJ2SYlfPuCth":"dmFsdWVfZm9yXzhCZE03SHoxcEoyU1lsZlB1Q3Ro","8JKaAygGixSarj0kkrGK":"dmFsdWVfZm9yXzhKS2FBeWdHaXhTYXJqMGtrckdL","9IdeyRyKSV1sCUSnAo6F":"dmFsdWVfZm9yXzlJZGV5UnlLU1Yxc0NVU25BbzZG","9O07WcUz7z398JsYqK0r":"dmFsdWVfZm9yXzlPMDdXY1V6N3ozOThKc1lxSzBy","9QqRtmHDErfX0ZECOfCF":"dmFsdWVfZm9yXzlRcVJ0bUhERXJmWDBaRUNPZkNG","9y5PWIQ4LX57uIeT1G3q":"dmFsdWVfZm9yXzl5NVBXSVE0TFg1N3VJZVQxRzNx","AWzcti08vsv2QfwMfR5b":"dmFsdWVfZm9yX0FXemN0aTA4dnN2MlFmd01mUjVi","C8py7RYDcnehWMbvJ6Qx":"dmFsdWVfZm9yX0M4cHk3UllEY25laFdNYnZKNlF4","CFepjYMANacdNrnWmRIy":"dmFsdWVfZm9yX0NGZXBqWU1BTmFjZE5ybldtUkl5","CKAzrXL0L7O9RSkTYV1o":"dmFsdWVfZm9yX0NLQXpyWEwwTDdPOVJTa1RZVjFv","DxcUoyoDOJTtrWN17WVz":"dmFsdWVfZm9yX0R4Y1VveW9ET0pUdHJXTjE3V1Z6","EFqsocQduyfkFBf2wlQR":"dmFsdWVfZm9yX0VGcXNvY1FkdXlma0ZCZjJ3bFFS","F8RHOsR3IgBLDU18kbx6":"dmFsdWVfZm9yX0Y4UkhPc1IzSWdCTERVMThrYng2","FU6sf8MeyWpcI6eEP9Lo":"dmFsdWVfZm9yX0ZVNnNmOE1leVdwY0k2ZUVQOUxv","FYfCAWVbx8P0eo86ns9C":"dmFsdWVfZm9yX0ZZZkNBV1ZieDhQMGVvODZuczlD","FzUmQyWgaDKZoZurOJYt":"dmFsdWVfZm9yX0Z6VW1ReVdnYURLWm9adXJPSll0","G709H7D4HJJJbXVDPZUL":"dmFsdWVfZm9yX0c3MDlIN0Q0SEpKSmJYVkRQWlVM","GKJfAoSKGda5neq1JMH8":"dmFsdWVfZm9yX0dLSmZBb1NLR2RhNW5lcTFKTUg4","GuCYPBLjbfCLsLhReX61":"dmFsdWVfZm9yX0d1Q1lQQkxqYmZDTHNMaFJlWDYx","ItoKojOJHipAINrm1Y2m":"dmFsdWVfZm9yX0l0b0tvak9KSGlwQUlOcm0xWTJt","IvFiv00qvcpol9En197O":"dmFsdWVfZm9yX0l2Rml2MDBxdmNwb2w5RW4xOTdP","JHGTNtrG54UQyDoOUIhu":"dmFsdWVfZm9yX0pIR1ROdHJHNTRVUXlEb09VSWh1","KcQo19Nfb3yEqvk8gkWM":"dmFsdWVfZm9yX0tjUW8xOU5mYjN5RXF2azhna1dN","KmVFtdL2IhdHPAxPkSLE":"dmFsdWVfZm9yX0ttVkZ0ZEwySWhkSFBBeFBrU0xF","Ky2ZbWhGqVxio9XyhFaW":"dmFsdWVfZm9yX0t5MlpiV2hHcVZ4aW85WHloRmFX","LJQCoRjX7vfBRfDbQt8H":"dmFsdWVfZm9yX0xKUUNvUmpYN3ZmQlJmRGJRdDhI","Lx6VgbJOwCkBhNPbNPLK":"dmFsdWVfZm9yX0x4NlZnYkpPd0NrQmhOUGJOUExL","NCT3ZrF4GXKJx7GdOFOS":"dmFsdWVfZm9yX05DVDNackY0R1hLSng3R2RPRk9T","NIhTWrJ9s08uPMlnZVvQ":"dmFsdWVfZm9yX05JaFRXcko5czA4dVBNbG5aVnZR","OpRVSOIfHqORPsVXDKDV":"dmFsdWVfZm9yX09wUlZTT0lmSHFPUlBzVlhES0RW","PRuR6E7oEzt6P72aFbwZ":"dmFsdWVfZm9yX1BSdVI2RTdvRXp0NlA3MmFGYnda","Pno8t6KUVTJX3SOv8oJ9":"dmFsdWVfZm9yX1Bubzh0NktVVlRKWDNTT3Y4b0o5","QJyRHy5l8wxWsD3RAD9V":"dmFsdWVfZm9yX1FKeVJIeTVsOHd4V3NEM1JBRDlW","QuYf3ZXEszekWIr5S3Xa":"dmFsdWVfZm9yX1F1WWYzWlhFc3pla1dJcjVTM1hh","R8Ks6fJC0z0zroF21Z8k":"dmFsdWVfZm9yX1I4S3M2ZkpDMHowenJvRjIxWjhr","Rayla0tIchrWOL1ITfwD":"dmFsdWVfZm9yX1JheWxhMHRJY2hyV09MMUlUZndE","RgbjnxLsdv5TJkBgFNkc":"dmFsdWVfZm9yX1JnYmpueExzZHY1VEprQmdGTmtj","S5xW6udsDnBWx94ACQZY":"dmFsdWVfZm9yX1M1eFc2dWRzRG5CV3g5NEFDUVpZ","SYlcZLCYlJ6FyeYWDwKa":"dmFsdWVfZm9yX1NZbGNaTENZbEo2RnllWVdEd0th","SxdN0sbN2XnyXp8D759v":"dmFsdWVfZm9yX1N4ZE4wc2JOMlhueVhwOEQ3NTl2","TuVHcGSvaBkAZfoJdOnG":"dmFsdWVfZm9yX1R1VkhjR1N2YUJrQVpmb0pkT25H","VI2rkW5FbVgrFBuIjl9O":"dmFsdWVfZm9yX1ZJMnJrVzVGYlZnckZCdUlqbDlP","Vr5WHngEaaObl7vm4sW3":"dmFsdWVfZm9yX1ZyNVdIbmdFYWFPYmw3dm00c1cz","VtGTl4UDazfSNCCvif6f":"dmFsdWVfZm9yX1Z0R1RsNFVEYXpmU05DQ3ZpZjZm","Xz7nRWjIvT1qAeESWKc0":"dmFsdWVfZm9yX1h6N25SV2pJdlQxcUFlRVNXS2Mw","YVheNg6NNEAHaR8FW8tw":"dmFsdWVfZm9yX1lWaGVOZzZOTkVBSGFSOEZXOHR3","Ye6pzTgH56pRHVTwyqKI":"dmFsdWVfZm9yX1llNnB6VGdINTZwUkhWVHd5cUtJ","ZI276YyLEe4Rk3J2UcyW":"dmFsdWVfZm9yX1pJMjc2WXlMRWU0UmszSjJVY3lX","a4DtclQeVQYKKIKDyKTR":"dmFsdWVfZm9yX2E0RHRjbFFlVlFZS0tJS0R5S1RS","bClmdTPI7TMtXM9bqxNl":"dmFsdWVfZm9yX2JDbG1kVFBJN1RNdFhNOWJxeE5s","cv7glCmz2W2XW3iNCbsH":"dmFsdWVfZm9yX2N2N2dsQ216MlcyWFczaU5DYnNI","dciULheIoid3zGTssPwb":"dmFsdWVfZm9yX2RjaVVMaGVJb2lkM3pHVHNzUHdi","dhTD65H2O5QEuQ75M824":"dmFsdWVfZm9yX2RoVEQ2NUgyTzVRRXVRNzVNODI0","f8KV6BlY5OOV5XKBSJqj":"dmFsdWVfZm9yX2Y4S1Y2QmxZNU9PVjVYS0JTSnFq","j44mffiWtKKAbjTGDP1H":"dmFsdWVfZm9yX2o0NG1mZmlXdEtLQWJqVEdEUDFI","jmh7jfqajzc9wn7MoZZI":"dmFsdWVfZm9yX2ptaDdqZnFhanpjOXduN01vWlpJ","kFAMcj9mnIVSoFp48KM2":"dmFsdWVfZm9yX2tGQU1jajltbklWU29GcDQ4S00y","kIw0ZPSL4KihdRAas0zq":"dmFsdWVfZm9yX2tJdzBaUFNMNEtpaGRSQWFzMHpx","lPyUwFXBI6J8cB8lcdtl":"dmFsdWVfZm9yX2xQeVV3RlhCSTZKOGNCOGxjZHRs","lY99bJIMLBiGRPvauYTU":"dmFsdWVfZm9yX2xZOTliSklNTEJpR1JQdmF1WVRV","ljU4BMrrhRki21hxdmDL":"dmFsdWVfZm9yX2xqVTRCTXJyaFJraTIxaHhkbURM","m2rY4gnfexhTjdi0YJ2G":"dmFsdWVfZm9yX20yclk0Z25mZXhoVGpkaTBZSjJH","mP79bvOH5rGyHUpzRAA4":"dmFsdWVfZm9yX21QNzlidk9INXJHeUhVcHpSQUE0","mbSxdFQaN846izRL2JlV":"dmFsdWVfZm9yX21iU3hkRlFhTjg0Nml6UkwySmxW","ndDq0eYqiHU9M5X6b9Fb":"dmFsdWVfZm9yX25kRHEwZVlxaUhVOU01WDZiOUZi","nifBmcrEBMthjU1J7iwT":"dmFsdWVfZm9yX25pZkJtY3JFQk10aGpVMUo3aXdU","nrjiAMzDSpjb3RhfQ3ue":"dmFsdWVfZm9yX25yamlBTXpEU3BqYjNSaGZRM3Vl","o7bgcuSm3Yp0znzVb0dL":"dmFsdWVfZm9yX283YmdjdVNtM1lwMHpuelZiMGRM","o8qHdoSCuhPpIunvfAko":"dmFsdWVfZm9yX284cUhkb1NDdWhQcEl1bnZmQWtv","pUtl9tzdxo48ty0Fm1PI":"dmFsdWVfZm9yX3BVdGw5dHpkeG80OHR5MEZtMVBJ","qGk4TDuyYT4pXLuri6Is":"dmFsdWVfZm9yX3FHazRURHV5WVQ0cFhMdXJpNklz","qxIGv4vSKxCYSdVSiaCy":"dmFsdWVfZm9yX3F4SUd2NHZTS3hDWVNkVlNpYUN5","rT6zKAZFy4kqhVN3jTEM":"dmFsdWVfZm9yX3JUNnpLQVpGeTRrcWhWTjNqVEVN","s9JGkIs6E1JUA6cvGVVl":"dmFsdWVfZm9yX3M5SkdrSXM2RTFKVUE2Y3ZHVlZs","tXuQmyBvsC7HsIZvtDmY":"dmFsdWVfZm9yX3RYdVFteUJ2c0M3SHNJWnZ0RG1Z","vPt18lCpqoGEh01BhIS6":"dmFsdWVfZm9yX3ZQdDE4bENwcW9HRWgwMUJoSVM2","wc6xrExjothp2EjS6gtA":"dmFsdWVfZm9yX3djNnhyRXhqb3RocDJFalM2Z3RB","wdTIepIoREYK50sW555A":"dmFsdWVfZm9yX3dkVEllcElvUkVZSzUwc1c1NTVB","wthUNphkJd9ZCI162pAw":"dmFsdWVfZm9yX3d0aFVOcGhrSmQ5WkNJMTYycEF3","wwo0tAbsGOtfq5Svfc5x":"dmFsdWVfZm9yX3d3bzB0QWJzR090ZnE1U3ZmYzV4","xGkk2jQi3WgFU9AnxWBN":"dmFsdWVfZm9yX3hHa2syalFpM1dnRlU5QW54V0JO","xc8OyQWXdpzNdS78kJje":"dmFsdWVfZm9yX3hjOE95UVdYZHB6TmRTNzhrSmpl","y3PmJTcoEv429Y1bdpMm":"dmFsdWVfZm9yX3kzUG1KVGNvRXY0MjlZMWJkcE1t","ywCzRLVWFnx5oDEzEI8I":"dmFsdWVfZm9yX3l3Q3pSTFZXRm54NW9ERXpFSThJ","zOEEtX7nrhQdsUZ9dY45":"dmFsdWVfZm9yX3pPRUV0WDducmhRZHNVWjlkWTQ1","zmLAuzJMA5xQQN7sp9ce":"dmFsdWVfZm9yX3ptTEF1ekpNQTV4UVFON3NwOWNl","zpFHYEHvXeBB6ubOGMsI":"dmFsdWVfZm9yX3pwRkhZRUh2WGVCQjZ1Yk9HTXNJ","zwddEVYPRxYO5QcUnEsQ":"dmFsdWVfZm9yX3p3ZGRFVllQUnhZTzVRY1VuRXNR"},"Key":"\ufffd\ufffd\ufffd\ufffd"}
