--- fuzz/store/internal/proofs/createnonmembershipproof/corpus/5.txt
@@ -1,1 +1,0 @@
-{"Data":{"03BrFVEi28Unudk2IxmE":"dmFsdWVfZm9yXzAzQnJGVkVpMjhVbnVkazJJeG1F","0LN9FvvVcb8ltW7U8KiP":"dmFsdWVfZm9yXzBMTjlGdnZWY2I4bHRXN1U4S2lQ","0x6hTUyZt7WToBYsckoe":"dmFsdWVfZm9yXzB4NmhUVXladDdXVG9CWXNja29l","1CoJsErrCtnU9hrKbyK4":"dmFsdWVfZm9yXzFDb0pzRXJyQ3RuVTlocktieUs0","1Ld3ro296JtlFlXlSkzG":"dmFsdWVfZm9yXzFMZDNybzI5Nkp0bEZsWGxTa3pH","1sfYBRueXunNrnhvnLTV":"dmFsdWVfZm9yXzFzZllCUnVlWHVuTnJuaHZuTFRW","4piNQeRO018Prwzl8m1x":"dmFsdWVfZm9yXzRwaU5RZVJPMDE4UHJ3emw4bTF4","4qgeQu0qFxng7KpnisJR":"dmFsdWVfZm9yXzRxZ2VRdTBxRnhuZzdLcG5pc0pS","6Smj5QF8OqTazRUIpN3O":"dmFsdWVfZm9yXzZTbWo1UUY4T3FUYXpSVUlwTjNP","6uT6QSVgrY4AclZ3tg5d":"dmFsdWVfZm9yXzZ1VDZRU1Znclk0QWNsWjN0ZzVk","7EAHwGMnP76YsUfOurBe":"dmFsdWVfZm9yXzdFQUh3R01uUDc2WXNVZk91ckJl","7rhnhlsIjwpLTI4UL6dK":"dmFsdWVfZm9yXzdyaG5obHNJandwTFRJNFVMNmRL","83LL2Tk578DKW6jXJTs4":"dmFsdWVfZm9yXzgzTEwyVGs1NzhES1c2alhKVHM0","9A6NzRnn3NuXuQexdFsX":"dmFsdWVfZm9yXzlBNk56Um5uM051WHVRZXhkRnNY","9UhVRpy3fAkLosvTEqex":"dmFsdWVfZm9yXzlVaFZScHkzZkFrTG9zdlRFcWV4","AQDx4zjdDeI0GcK5emnq":"dmFsdWVfZm9yX0FRRHg0empkRGVJMEdjSzVlbW5x","Bv4i1Y67wlfCcx8Xv9TQ":"dmFsdWVfZm9yX0J2NGkxWTY3d2xmQ2N4OFh2OVRR","Bv55PeBFxO9CENE1AJy1":"dmFsdWVfZm9yX0J2NTVQZUJGeE85Q0VORTFBSnkx","CWyxwzTnQ6IZlVSTKnlh":"dmFsdWVfZm9yX0NXeXh3elRuUTZJWmxWU1RLbmxo","CdpeprT7zLmk4FI5Zahk":"dmFsdWVfZm9yX0NkcGVwclQ3ekxtazRGSTVaYWhr","EUBAbEknDS2l9psJIveL":"dmFsdWVfZm9yX0VVQkFiRWtuRFMybDlwc0pJdmVM","F3AvTy7gQDx1vM7fV7WR":"dmFsdWVfZm9yX0YzQXZUeTdnUUR4MXZNN2ZWN1dS","F6GScWFdXPZOEeJvYJKZ":"dmFsdWVfZm9yX0Y2R1NjV0ZkWFBaT0VlSnZZSkta","FgwpK4WdP9iyLl1gQktd":"dmFsdWVfZm9yX0Znd3BLNFdkUDlpeUxsMWdRa3Rk","HVCQR900ifxBnGblUN2g":"dmFsdWVfZm9yX0hWQ1FSOTAwaWZ4Qm5HYmxVTjJn","I0WfM4JkhSLpYT2fjEPw":"dmFsdWVfZm9yX0kwV2ZNNEpraFNMcFlUMmZqRVB3","J939leaYFHC0pdHvN3pJ":"dmFsdWVfZm9yX0o5MzlsZWFZRkhDMHBkSHZOM3BK","KC008u2yAT8hlgAWep6S":"dmFsdWVfZm9yX0tDMDA4dTJ5QVQ4aGxnQVdlcDZT","LCyhuJNtTb1XDArJo4dt":"dmFsdWVfZm9yX0xDeWh1Sk50VGIxWERBckpvNGR0","LtIZ8YZqrqP9RGW84OEy":"dmFsdWVfZm9yX0x0SVo4WVpxcnFQOVJHVzg0T0V5","MOb8fxFvzi53bQxxz2Wb":"dmFsdWVfZm9yX01PYjhmeEZ2emk1M2JReHh6Mldi","MiX82TdylEvMDs743AIO":"dmFsdWVfZm9yX01pWDgyVGR5bEV2TURzNzQzQUlP","MzFykpOsDDZs5Xxgrq15":"dmFsdWVfZm9yX016RnlrcE9zRERaczVYeGdycTE1","NH1rQB1mAWpBjZHqGlaL":"dmFsdWVfZm9yX05IMXJRQjFtQVdwQmpaSHFHbGFM","NTvi9ZXQZiFefo4T3ImM":"dmFsdWVfZm9yX05Udmk5WlhRWmlGZWZvNFQzSW1N","Nzn1AIlu1HlcimdDmREf":"dmFsdWVfZm9yX056bjFBSWx1MUhsY2ltZERtUkVm","O02LQ3M5YpZPKEdqVGGp":"dmFsdWVfZm9yX08wMkxRM001WXBaUEtFZHFWR0dw","O67Op1YTcwVNeEYjc2A5":"dmFsdWVfZm9yX082N09wMVlUY3dWTmVFWWpjMkE1","PCZfl9WryU6iQDCotaB3":"dmFsdWVfZm9yX1BDWmZsOVdyeVU2aVFEQ290YUIz","PsSOyMSzoyxi9hOa7Ulc":"dmFsdWVfZm9yX1BzU095TVN6b3l4aTloT2E3VWxj","QObQpkiJ4acxHlc78rOT":"dmFsdWVfZm9yX1FPYlFwa2lKNGFjeEhsYzc4ck9U","RJp8p72QcShiaUqckY5j":"dmFsdWVfZm9yX1JKcDhwNzJRY1NoaWFVcWNrWTVq","RYAeRQIDLnbEIHFTaNnP":"dmFsdWVfZm9yX1JZQWVSUUlETG5iRUlIRlRhTm5Q","RvzhnYruwlyjRSBwp1qG":"dmFsdWVfZm9yX1J2emhuWXJ1d2x5alJTQndwMXFH","SeNPAhNcXJ5F87PGaIrS":"dmFsdWVfZm9yX1NlTlBBaE5jWEo1Rjg3UEdhSXJT","UXosaMtVA1Rd8Gbm9p8s":"dmFsdWVfZm9yX1VYb3NhTXRWQTFSZDhHYm05cDhz","V6RtiP89bFOkIyNDfhMa":"dmFsdWVfZm9yX1Y2UnRpUDg5YkZPa0l5TkRmaE1h","VfhQQsWe2s5XFScYGei5":"dmFsdWVfZm9yX1ZmaFFRc1dlMnM1WEZTY1lHZWk1","VsZM0SFbRJ86F1i8s5SE":"dmFsdWVfZm9yX1ZzWk0wU0ZiUko4NkYxaThzNVNF","WKF1GHTwD9vryWmfdkvG":"dmFsdWVfZm9yX1dLRjFHSFR3RDl2cnlXbWZka3ZH","WQNSEnqPymowgBxCr9Vq":"dmFsdWVfZm9yX1dRTlNFbnFQeW1vd2dCeENyOVZx","WR8HTi4CZNgPxfD3jfFL":"dmFsdWVfZm9yX1dSOEhUaTRDWk5nUHhmRDNqZkZM","XQw5XblJxMKB1WuEfdgc":"dmFsdWVfZm9yX1hRdzVYYmxKeE1LQjFXdUVmZGdj","Xo886r8lYaBbaFQZpaLW":"dmFsdWVfZm9yX1hvODg2cjhsWWFCYmFGUVpwYUxX","YL9aES4DWzXkI20cLlFy":"dmFsdWVfZm9yX1lMOWFFUzREV3pYa0kyMGNMbEZ5","YRCE7ulm43AbeRFBBh8F":"dmFsdWVfZm9yX1lSQ0U3dWxtNDNBYmVSRkJCaDhG","b7W7EVR8dbEjahilLHmA":"dmFsdWVfZm9yX2I3VzdFVlI4ZGJFamFoaWxMSG1B","bH63WRPZIAeDOiL6Ocep":"dmFsdWVfZm9yX2JINjNXUlBaSUFlRE9pTDZPY2Vw","cMIZds4RAVRGO2epJTQ3":"dmFsdWVfZm9yX2NNSVpkczRSQVZSR08yZXBKVFEz","ccMVwgW5Q6AVxI7DRB2G":"dmFsdWVfZm9yX2NjTVZ3Z1c1UTZBVnhJN0RSQjJH","cdOhK2jfAXIbyxToFTG4":"dmFsdWVfZm9yX2NkT2hLMmpmQVhJYnl4VG9GVEc0","dezQbLFfnXzYofRm0Ica":"dmFsdWVfZm9yX2RlelFiTEZmblh6WW9mUm0wSWNh","f5lQCaUzHhlTsZzLtq3y":"dmFsdWVfZm9yX2Y1bFFDYVV6SGhsVHNaekx0cTN5","f92iQN7kqWoQK5cUjRNi":"dmFsdWVfZm9yX2Y5MmlRTjdrcVdvUUs1Y1VqUk5p","fWDCVwjkVyuW3uVrBybP":"dmFsdWVfZm9yX2ZXRENWd2prVnl1VzN1VnJCeWJQ","fsZbb5GPqNXm9fRzGwMs":"dmFsdWVfZm9yX2ZzWmJiNUdQcU5YbTlmUnpHd01z","he52FzWxlG7kB42v5UZ7":"dmFsdWVfZm9yX2hlNTJGeld4bEc3a0I0MnY1VVo3","heO5zHpdMGArK9qbM5UY":"dmFsdWVfZm9yX2hlTzV6SHBkTUdBcks5cWJNNVVZ","hm7JVuzP7NkoYuVSG5u4":"dmFsdWVfZm9yX2htN0pWdXpQN05rb1l1VlNHNXU0","ijGfjKoQ8x8yOUKeCdIj":"dmFsdWVfZm9yX2lqR2ZqS29ROHg4eU9VS2VDZElq","jhlEa1Rog6NIe9fdtdec":"dmFsdWVfZm9yX2pobEVhMVJvZzZOSWU5ZmR0ZGVj","jjtWHoSX4KGbe3g0GTLJ":"dmFsdWVfZm9yX2pqdFdIb1NYNEtHYmUzZzBHVExK","klQX6K519FpBfFma1dnn":"dmFsdWVfZm9yX2tsUVg2SzUxOUZwQmZGbWExZG5u","l3aTeXFWjOBCQTSQAj2n":"dmFsdWVfZm9yX2wzYVRlWEZXak9CQ1FUU1FBajJu","lFQj1nQtJUPlbE1qX4de":"dmFsdWVfZm9yX2xGUWoxblF0SlVQbGJFMXFYNGRl","mJSD2pS9YIef5nhbzg8T":"dmFsdWVfZm9yX21KU0QycFM5WUllZjVuaGJ6ZzhU","mgvficMgmcRnU1RUj1aO":"dmFsdWVfZm9yX21ndmZpY01nbWNSblUxUlVqMWFP","ndBEm9MxxOkdMTiNTUiW":"dmFsdWVfZm9yX25kQkVtOU14eE9rZE1UaU5UVWlX","oJfqLPggzgOtSRJaa0z4":"dmFsdWVfZm9yX29KZnFMUGdnemdPdFNSSmFhMHo0","oTubDMI9CNj1kAAhdVaz":"dmFsdWVfZm9yX29UdWJETUk5Q05qMWtBQWhkVmF6","ok6VXHjVcSHmoLypEca7":"dmFsdWVfZm9yX29rNlZYSGpWY1NIbW9MeXBFY2E3","pgd7CNNC525puJ5WcnW8":"dmFsdWVfZm9yX3BnZDdDTk5DNTI1cHVKNVdjblc4","qLDyqTfQBZcNrXcA9GpG":"dmFsdWVfZm9yX3FMRHlxVGZRQlpjTnJYY0E5R3BH","qMsRqyaa0iCUmCQggEiy":"dmFsdWVfZm9yX3FNc1JxeWFhMGlDVW1DUWdnRWl5","sDCB0P36DvzXgNfnFZ1b":"dmFsdWVfZm9yX3NEQ0IwUDM2RHZ6WGdOZm5GWjFi","scugxgT9whqtySMqiouu":"dmFsdWVfZm9yX3NjdWd4Z1Q5d2hxdHlTTXFpb3V1","skdkKgpPtHFl9jINQOwL":"dmFsdWVfZm9yX3NrZGtLZ3BQdEhGbDlqSU5RT3dM","tP791dk4cx4xBaQ8LW6e":"dmFsdWVfZm9yX3RQNzkxZGs0Y3g0eEJhUThMVzZl","tb4HHrU7At9NdH4c4jOU":"dmFsdWVfZm9yX3RiNEhIclU3QXQ5TmRINGM0ak9V","tvDpt0n1gZ0dT4dVyqZ3":"dmFsdWVfZm9yX3R2RHB0MG4xZ1owZFQ0ZFZ5cVoz","uzYjMMYQA4b0rzSKQTY5":"dmFsdWVfZm9yX3V6WWpNTVlRQTRiMHJ6U0tRVFk1","v43eyZTTrJIAp267EtEF":"dmFsdWVfZm9yX3Y0M2V5WlRUckpJQXAyNjdFdEVG","vr8PJcp0UnYbKkHnoMrd":"dmFsdWVfZm9yX3ZyOFBKY3AwVW5ZYktrSG5vTXJk","w5AKyHam5SwdQWzHsyHU":"dmFsdWVfZm9yX3c1QUt5SGFtNVN3ZFFXekhzeUhV","wjVjysCTvMZzmlP47MJX":"dmFsdWVfZm9yX3dqVmp5c0NUdk1aem1sUDQ3TUpY","x78H9cWz51nX7C0rSLAu":"dmFsdWVfZm9yX3g3OEg5Y1d6NTFuWDdDMHJTTEF1","xB5YVj4DA65qxmnxB5Jw":"dmFsdWVfZm9yX3hCNVlWajREQTY1cXhtbnhCNUp3","xD4j9y5OSHQRBRSlyDSY":"dmFsdWVfZm9yX3hENGo5eTVPU0hRUkJSU2x5RFNZ","yggUHsh9zV6EupW0KKSc":"dmFsdWVfZm9yX3lnZ1VIc2g5elY2RXVwVzBLS1Nj","z8iKIXtQl2oa43CmAAJV":"dmFsdWVfZm9yX3o4aUtJWHRRbDJvYTQzQ21BQUpW"},"Key":"\u0001\u0001\u0001\u0001"}
