+++ fuzz/store/internal_/proofs/createnonmembershipproof/corpus/6.txt
@@ -1,0 +1,1 @@
+{"Data":{"04qwRYwxXKaJz98LRbLg":"dmFsdWVfZm9yXzA0cXdSWXd4WEthSno5OExSYkxn","0UDrAZ7FluCiU06dkCLK":"dmFsdWVfZm9yXzBVRHJBWjdGbHVDaVUwNmRrQ0xL","0YXnw9SeCjx8Hd7cbMRH":"dmFsdWVfZm9yXzBZWG53OVNlQ2p4OEhkN2NiTVJI","0hRrmtkynIYinuAMJiEr":"dmFsdWVfZm9yXzBoUnJtdGt5bklZaW51QU1KaUVy","13lT6DEYiXVUeqEuLyNf":"dmFsdWVfZm9yXzEzbFQ2REVZaVhWVWVxRXVMeU5m","1XWd1FSfC0JrBNsOeIlY":"dmFsdWVfZm9yXzFYV2QxRlNmQzBKckJOc09lSWxZ","1aOYstzJZJpByGUmdFBy":"dmFsdWVfZm9yXzFhT1lzdHpKWkpwQnlHVW1kRkJ5","2qYIgff31knQyoxAouu2":"dmFsdWVfZm9yXzJxWUlnZmYzMWtuUXlveEFvdXUy","31x0xgP95JYRFMxbKq9V":"dmFsdWVfZm9yXzMxeDB4Z1A5NUpZUkZNeGJLcTlW","3V9a9nt2SyWO3ZeTv3W6":"dmFsdWVfZm9yXzNWOWE5bnQyU3lXTzNaZVR2M1c2","3ZnMCbOnZzA4OzCPfDdP":"dmFsdWVfZm9yXzNabk1DYk9uWnpBNE96Q1BmRGRQ","3ZnPz5dpnMnqbEEbtenV":"dmFsdWVfZm9yXzNablB6NWRwbk1ucWJFRWJ0ZW5W","4fiUcz0sRD3ON1seKOtq":"dmFsdWVfZm9yXzRmaVVjejBzUkQzT04xc2VLT3Rx","6P1TehOgkWS0638Y555j":"dmFsdWVfZm9yXzZQMVRlaE9na1dTMDYzOFk1NTVq","7D6eTF7tUkoWzVWOKuPa":"dmFsdWVfZm9yXzdENmVURjd0VWtvV3pWV09LdVBh","7PW0ocVNTN6GauMbpCik":"dmFsdWVfZm9yXzdQVzBvY1ZOVE42R2F1TWJwQ2lr","7mj1wHgATRprf3Jf9Lx3":"dmFsdWVfZm9yXzdtajF3SGdBVFJwcmYzSmY5THgz","9ne6sf1bB1Pnyceb8spU":"dmFsdWVfZm9yXzluZTZzZjFiQjFQbnljZWI4c3BV","C50ji6luu46NxLovODND":"dmFsdWVfZm9yX0M1MGppNmx1dTQ2TnhMb3ZPRE5E","CBlziF0GsUwSiC9o9HYj":"dmFsdWVfZm9yX0NCbHppRjBHc1V3U2lDOW85SFlq","CDMyHU6S2hkOL8F6YOZP":"dmFsdWVfZm9yX0NETXlIVTZTMmhrT0w4RjZZT1pQ","CW2Tw814pNfSL5PztoBL":"dmFsdWVfZm9yX0NXMlR3ODE0cE5mU0w1UHp0b0JM","DHaQam3FDxttkt3s0MAB":"dmFsdWVfZm9yX0RIYVFhbTNGRHh0dGt0M3MwTUFC","Egcl7UuDe87RF1VPnfiP":"dmFsdWVfZm9yX0VnY2w3VXVEZTg3UkYxVlBuZmlQ","EoHxORLqubB4FbQtKW1e":"dmFsdWVfZm9yX0VvSHhPUkxxdWJCNEZiUXRLVzFl","Eyw9XhuSveoa7JJKBcVB":"dmFsdWVfZm9yX0V5dzlYaHVTdmVvYTdKSktCY1ZC","FFXv1IaaijLHwz7QUQKk":"dmFsdWVfZm9yX0ZGWHYxSWFhaWpMSHd6N1FVUUtr","IVrk6YUjVMcY0kDtjwwb":"dmFsdWVfZm9yX0lWcms2WVVqVk1jWTBrRHRqd3di","JcbKs94H4un6pMua0yDE":"dmFsdWVfZm9yX0pjYktzOTRINHVuNnBNdWEweURF","JfCDsSMu2UfKCJaF9YKS":"dmFsdWVfZm9yX0pmQ0RzU011MlVmS0NKYUY5WUtT","KQG6me2ki5q8B6mnfm5N":"dmFsdWVfZm9yX0tRRzZtZTJraTVxOEI2bW5mbTVO","KpldKScR5xlPGJKFC4Mq":"dmFsdWVfZm9yX0twbGRLU2NSNXhsUEdKS0ZDNE1x","LtuNnriszrFsMkqdTArd":"dmFsdWVfZm9yX0x0dU5ucmlzenJGc01rcWRUQXJk","LwxigICYVh5RWEkGz1nU":"dmFsdWVfZm9yX0x3eGlnSUNZVmg1UldFa0d6MW5V","Mx6sdqJq4IjAq6z1h2ol":"dmFsdWVfZm9yX014NnNkcUpxNElqQXE2ejFoMm9s","NKuMOYlLbzv8Hn7eY1eS":"dmFsdWVfZm9yX05LdU1PWWxMYnp2OEhuN2VZMWVT","Nz1aO08HgDNbKKRXqxm5":"dmFsdWVfZm9yX056MWFPMDhIZ0ROYktLUlhxeG01","O4c526pJCeD14gsg0sOK":"dmFsdWVfZm9yX080YzUyNnBKQ2VEMTRnc2cwc09L","Ol7yaJdlOGHP1Z2HpwaO":"dmFsdWVfZm9yX09sN3lhSmRsT0dIUDFaMkhwd2FP","OlBnUWMNpB7wRDvZXHDB":"dmFsdWVfZm9yX09sQm5VV01OcEI3d1JEdlpYSERC","PLvVF2uhhefyY1MSxwWs":"dmFsdWVfZm9yX1BMdlZGMnVoaGVmeVkxTVN4d1dz","R1lHxYmkg6w22yGzqTAx":"dmFsdWVfZm9yX1IxbEh4WW1rZzZ3MjJ5R3pxVEF4","R5c55sIupyrsCPw9Ttjg":"dmFsdWVfZm9yX1I1YzU1c0l1cHlyc0NQdzlUdGpn","R5nNWvJNxBXUy116dgJu":"dmFsdWVfZm9yX1I1bk5XdkpOeEJYVXkxMTZkZ0p1","RMmCgUcXPBGyzkIlP6l1":"dmFsdWVfZm9yX1JNbUNnVWNYUEJHeXprSWxQNmwx","RadFuKHkouO3FQdBrOaY":"dmFsdWVfZm9yX1JhZEZ1S0hrb3VPM0ZRZEJyT2FZ","T1j8XIMEWhyuIBIf6uA3":"dmFsdWVfZm9yX1QxajhYSU1FV2h5dUlCSWY2dUEz","U3dp7AnuvaQbuYzlgkDX":"dmFsdWVfZm9yX1UzZHA3QW51dmFRYnVZemxna0RY","U9L6UtmatRzjgDEhJU3h":"dmFsdWVfZm9yX1U5TDZVdG1hdFJ6amdERWhKVTNo","UWJBTibjsRV1w5uBGtYt":"dmFsdWVfZm9yX1VXSkJUaWJqc1JWMXc1dUJHdFl0","UfZb5hMEkoYsyXDhgSMh":"dmFsdWVfZm9yX1VmWmI1aE1Fa29Zc3lYRGhnU01o","Ws6nTnCbBINCTDlmcU6F":"dmFsdWVfZm9yX1dzNm5UbkNiQklOQ1REbG1jVTZG","Xw15sTCU07Kfkt1oWBtP":"dmFsdWVfZm9yX1h3MTVzVENVMDdLZmt0MW9XQnRQ","Ygtb3iHt0PEGXgv6JuKh":"dmFsdWVfZm9yX1lndGIzaUh0MFBFR1hndjZKdUto","ZNULndbkbcN0LDjCBmoE":"dmFsdWVfZm9yX1pOVUxuZGJrYmNOMExEakNCbW9F","ZmK2xBri2UBTdIGMH0AS":"dmFsdWVfZm9yX1ptSzJ4QnJpMlVCVGRJR01IMEFT","ZnecZRXeE4bJX3unti2D":"dmFsdWVfZm9yX1puZWNaUlhlRTRiSlgzdW50aTJE","bv5BuAnpqXMtPjyoznp3":"dmFsdWVfZm9yX2J2NUJ1QW5wcVhNdFBqeW96bnAz","cdFO68DZ7osdfKAte2qm":"dmFsdWVfZm9yX2NkRk82OERaN29zZGZLQXRlMnFt","dYwzxoNpqTAKgZqqIRfu":"dmFsdWVfZm9yX2RZd3p4b05wcVRBS2dacXFJUmZ1","dwvd5eQwRdckw0mlzv23":"dmFsdWVfZm9yX2R3dmQ1ZVF3UmRja3cwbWx6djIz","eBe1dMSOGyEIJSx59UXR":"dmFsdWVfZm9yX2VCZTFkTVNPR3lFSUpTeDU5VVhS","eSDw4jv9FJrPZ1qmMTgD":"dmFsdWVfZm9yX2VTRHc0anY5RkpyUFoxcW1NVGdE","fIOYaxEiq0GPL0pSWF2f":"dmFsdWVfZm9yX2ZJT1lheEVpcTBHUEwwcFNXRjJm","gS98bQWLBQAmb2IJxguu":"dmFsdWVfZm9yX2dTOThiUVdMQlFBbWIySUp4Z3V1","h1VeZ0lOh5MVsJOzOnzO":"dmFsdWVfZm9yX2gxVmVaMGxPaDVNVnNKT3pPbnpP","hB9lNiaLiqbKiuw40Llt":"dmFsdWVfZm9yX2hCOWxOaWFMaXFiS2l1dzQwTGx0","hTr6c09Crf7FTKO27Khl":"dmFsdWVfZm9yX2hUcjZjMDlDcmY3RlRLTzI3S2hs","hcglD2wjsnSNIISo2V0t":"dmFsdWVfZm9yX2hjZ2xEMndqc25TTklJU28yVjB0","iVmEsqmXPKqfMHnOykTw":"dmFsdWVfZm9yX2lWbUVzcW1YUEtxZk1Ibk95a1R3","ifPN45KzubSnAzvvwWqK":"dmFsdWVfZm9yX2lmUE40NUt6dWJTbkF6dnZ3V3FL","kuKgG26zmdwed5RUlLjV":"dmFsdWVfZm9yX2t1S2dHMjZ6bWR3ZWQ1UlVsTGpW","mBTnC9Qy5dCTuKGuwLjF":"dmFsdWVfZm9yX21CVG5DOVF5NWRDVHVLR3V3TGpG","mdbIPdti81pd8ie8SRoa":"dmFsdWVfZm9yX21kYklQZHRpODFwZDhpZThTUm9h","nS55cQ0dPuk3mQuNZ2Kw":"dmFsdWVfZm9yX25TNTVjUTBkUHVrM21RdU5aMkt3","oIScx4kQtXrmG1VUkD9k":"dmFsdWVfZm9yX29JU2N4NGtRdFhybUcxVlVrRDlr","oLGHHHIWxOxPeMKT9NXr":"dmFsdWVfZm9yX29MR0hISElXeE94UGVNS1Q5Tlhy","pEzoNVNYSSW7YhmchFeu":"dmFsdWVfZm9yX3BFem9OVk5ZU1NXN1lobWNoRmV1","pPQFimod6ZcjIEGiyRs8":"dmFsdWVfZm9yX3BQUUZpbW9kNlpjaklFR2l5UnM4","pkG6oyz9Z23P92ZUFH4Y":"dmFsdWVfZm9yX3BrRzZveXo5WjIzUDkyWlVGSDRZ","qOexSw6LM7A63DdpajRl":"dmFsdWVfZm9yX3FPZXhTdzZMTTdBNjNEZHBhalJs","qazkByq28TSSblJVkSA3":"dmFsdWVfZm9yX3FhemtCeXEyOFRTU2JsSlZrU0Ez","qkkQSPViLl9DBgiBfvZp":"dmFsdWVfZm9yX3Fra1FTUFZpTGw5REJnaUJmdlpw","rindnEOB9PGPDh1yVw1N":"dmFsdWVfZm9yX3JpbmRuRU9COVBHUERoMXlWdzFO","ruUVZ4DkLNBbtky7DFsJ":"dmFsdWVfZm9yX3J1VVZaNERrTE5CYnRreTdERnNK","rzxwDqSKsibB3WwOOKHC":"dmFsdWVfZm9yX3J6eHdEcVNLc2liQjNXd09PS0hD","s5mZU35NjoPbSZc1oHPS":"dmFsdWVfZm9yX3M1bVpVMzVOam9QYlNaYzFvSFBT","tNzGorq8REsIvY6ATfgb":"dmFsdWVfZm9yX3ROekdvcnE4UkVzSXZZNkFUZmdi","taHnB8aTEzyDT4oonGVR":"dmFsdWVfZm9yX3RhSG5COGFURXp5RFQ0b29uR1ZS","uFqZe7SWKmICq451Zuic":"dmFsdWVfZm9yX3VGcVplN1NXS21JQ3E0NTFadWlj","vHoNuEftNnDRBEEjUzIf":"dmFsdWVfZm9yX3ZIb051RWZ0Tm5EUkJFRWpVeklm","wWoJmIT5GK8dw6ax2QkJ":"dmFsdWVfZm9yX3dXb0ptSVQ1R0s4ZHc2YXgyUWtK","wj1RtfUuHZeqGl9uSXVj":"dmFsdWVfZm9yX3dqMVJ0ZlV1SFplcUdsOXVTWFZq","wrGUzdCPOo9vrH0BQ4Y8":"dmFsdWVfZm9yX3dyR1V6ZENQT285dnJIMEJRNFk4","x8fPWc0L4OEYBckIOXV3":"dmFsdWVfZm9yX3g4ZlBXYzBMNE9FWUJja0lPWFYz","yFHixecgHoK86XPQvjSU":"dmFsdWVfZm9yX3lGSGl4ZWNnSG9LODZYUFF2alNV","yXGvfabhkeROJyxvez38":"dmFsdWVfZm9yX3lYR3ZmYWJoa2VST0p5eHZlejM4","ykrx09hx38JPA824KAqz":"dmFsdWVfZm9yX3lrcngwOWh4MzhKUEE4MjRLQXF6","zNPn1FaALSBqmx5P9L4R":"dmFsdWVfZm9yX3pOUG4xRmFBTFNCcW14NVA5TDRS","zmaWhOL7HSN1QoNW6uB9":"dmFsdWVfZm9yX3ptYVdoT0w3SFNOMVFvTlc2dUI5"},"Key":"OlBnUWMNpB7wRDvZXH\ufffd\ufffd"}
